fix path create course and move Seclect to components
