+++ frontend/src/App.tsx
@@ -13,7 +13,7 @@ import CreateCourse from "./pages/CreateCourse";
 import Register from "./pages/Register";
 import PrivateRoute from "./routes/PrivateRoute";
 import Verify from "./pages/Verify";
-// import NotFound from "./pages/NotFound";
+import NotFound from "./pages/NotFound";
 import MyCourses from "./pages/MyCourse";
 import Cookies from "js-cookie";
 import EditCourse from "./pages/EditCourse/EditCourse";
@@ -45,11 +45,11 @@ function App() {
                         <Route element={<PrivateRoute />}>
                             <Route path="/change-password" element={<ChangePassword />}></Route>
                             <Route path="/my-courses" element={<MyCourses />}></Route>
+                            <Route path="/create-course" element={<CreateCourse />}></Route>
+                            <Route path="/my-courses/edit/:course_id" element={<EditCourse />}></Route>
                         </Route>
                         <Route path="/my-profile" element={<MyProfile />}></Route>
                         <Route path="/profile/:id" element={<AuthorProfile />}></Route>
-                        <Route path="/create-course" element={<CreateCourse />}></Route>
-                        <Route path="/my-courses/edit/:course_id" element={<EditCourse />}></Route>
                         <Route path="/course-detail/:slug" element={<CourseDetail />}></Route>
                         <Route path="/forgot-password" element={<ForgotPassword />}></Route>
                         <Route path="/reset-password/:token" element={<ResetPassword />}></Route>
@@ -57,7 +57,7 @@ function App() {
                         <Route path="/register" element={<Register />}></Route>
                         <Route path="/verify-email/:token" element={<Verify />}></Route>
                         <Route path="/course-detail/:slug/watch" element={<WatchVideo />}></Route>
-                        {/* <Route path="/*" element={<NotFound />}></Route> */}
+                        <Route path="/*" element={<NotFound />}></Route>
                     </Routes>
                     <Footer />
                 </div>
